requirements.txt
absl-py==2.1.0
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coverage==7.6.1
deepface==0.0.85
exceptiongroup==1.3.0
filelock==3.13.1
fire==0.5.0
Flask==3.0.2
flatbuffers==23.5.26
gast==0.4.0
gdown==5.1.0
google-auth==2.28.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.62.0
gunicorn==21.2.0
h5py==3.10.0
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
keras==2.13.1
Keras-Preprocessing==1.1.2
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
mtcnn==0.1.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.0.3
pillow==10.2.0
pluggy==1.5.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
PySocks==1.7.1
pytest==8.3.5
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.14
rsa==4.9
six==1.16.0
soupsieve==2.5
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
tomli==2.3.0
tqdm==4.66.2
typing_extensions==4.13.2
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.17.0
